# api-testing-suite/docs/api_spec.yaml
openapi: 3.0.3
info:
  title: Payment Processing API
  description: |
    A mock payment processing API designed for integration testing.
    Supports payment creation, retrieval, listing with pagination,
    refunds, status checks, webhooks, and OAuth2 authentication flows.

    ## Authentication
    All protected endpoints require a valid API key passed via:
    - `Authorization: Bearer demo_key_...` header
    - `X-API-Key: demo_key_...` header

    ## Rate Limiting
    Requests are rate-limited per API key. Standard keys allow 100
    requests per minute. Rate limit headers are included in every response.

    ## Error Handling
    All errors follow a consistent JSON format with `type`, `message`,
    and `status` fields. Validation errors include a detailed `errors`
    array with per-field information.
  version: 2.4.0
  contact:
    name: API Engineering Team
    email: api-team@example.com
  license:
    name: MIT
    url: https://opensource.org/licenses/MIT
servers:
  - url: http://localhost:5000/v1
    description: Local development server
  - url: https://api.staging.example.com/v1
    description: Staging environment
...
paths:
...
  /webhooks/endpoints:
    post:
      tags: [Webhooks]
      summary: Register webhook endpoint
      description: Register a URL to receive webhook event notifications. Requires admin permissions.
      operationId: registerWebhook
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              required: [url]
              properties:
                url:
                  type: string
                  format: uri
                  description: HTTPS URL for webhook delivery
                events:
                  type: array
                  items:
                    type: string
                  description: Event types to subscribe to
      responses:
        "201":
          description: Webhook endpoint registered
        "422":
          description: Validation error
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ValidationError"

    get:
      tags: [Webhooks]
      summary: List webhook endpoints
      operationId: listWebhooks
      responses:
        "200":
          description: List of registered webhooks
  /webhooks/endpoints/{webhook_id}:
    delete:
      tags: [Webhooks]
      summary: Delete webhook endpoint
      operationId: deleteWebhook
      parameters:
        - name: webhook_id
          in: path
          required: true
          schema:
            type: string
      responses:
        "200":
          description: Webhook deleted
        "404":
          description: Webhook not found
  /webhooks/events:
    get:
      tags: [Webhooks]
      summary: List webhook events
      operationId: listWebhookEvents
      parameters:
        - name: type
          in: query
          schema:
            type: string
        - name: limit
          in: query
          schema:
            type: integer
            default: 20
            maximum: 100
      responses:
        "200":
          description: List of webhook events
  /webhooks/simulate:
    post:
      tags: [Webhooks]
      summary: Simulate webhook event
      description: Fire a test webhook event for development and testing.
      operationId: simulateWebhook
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              required: [event_type]
              properties:
                event_type:
                  type: string
                  enum:
                    - payment.created
                    - payment.completed
                    - payment.failed
                    - payment.cancelled
                    - refund.created
                    - refund.completed
                    - refund.failed
                payload:
                  type: object
      responses:
        "200":
          description: Webhook event simulated
...
components:
  securitySchemes:
    BearerAuth:
      type: http
      scheme: bearer
      description: API key passed as a Bearer token
    ApiKeyHeader:
      type: apiKey
      in: header
      name: X-API-Key
      description: API key passed in a custom header
  schemas:
...
    ValidationError:
      type: object
      properties:
        error:
          type: object
          properties:
            type:
              type: string
              enum: [validation_error]
            message:
              type: string
            status:
              type: integer
              enum: [422]
            errors:
              type: array
              items:
                type: object
                properties:
                  field:
                    type: string
                  code:
                    type: string
                  message:
                    type: string
security:
  - BearerAuth: []
  - ApiKeyHeader: []